backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
tensorflow
numpy
python-dotenv
ultralytics==8.3.228
opencv-python>=4.8.0
pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
albumentations>=1.3.0
matplotlib>=3.7.0
tensorboard>=2.14.0
opencv-python
reportlab
scikit-learn